# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Jumyste application",
        "title": "Jumyste App API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/resume/": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete the resume associated with a given user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resume"
                ],
                "summary": "Delete a resume by user ID",
                "responses": {
                    "200": {
                        "description": "Successfully deleted resume",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete resume",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/resume/manual": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create and save a resume using the provided JSON data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resume"
                ],
                "summary": "Create a resume from JSON data",
                "parameters": [
                    {
                        "description": "Resume data",
                        "name": "resume_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ResumeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resume saved successfully",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to save resume",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/resume/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload a resume file and process it",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resume"
                ],
                "summary": "Upload a resume",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Resume file",
                        "name": "resume",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ResumeResponse"
                        }
                    },
                    "400": {
                        "description": "Failed to retrieve resume file",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to process resume",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/resume/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the resume of a user by their user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resume"
                ],
                "summary": "Get a resume by user ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ResumeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Resume or user not found",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get resume",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the profile information of the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get user information",
                "responses": {
                    "200": {
                        "description": "User profile information",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_entity.UserResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Token is missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found - No user associated with the given ID",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error - Invalid user ID type",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update user details such as name, email, or profile picture",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update user information",
                "parameters": [
                    {
                        "description": "Fields to update",
                        "name": "updates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update user",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "jumyste-app-backend_internal_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid input"
                }
            }
        },
...
        "jumyste-app-backend_internal_dto.ResumeRequest": {
            "type": "object",
            "required": [
                "desired_position",
                "full_name",
                "skills"
            ],
            "properties": {
                "about": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "desired_position": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "jumyste-app-backend_internal_dto.ResumeResponse": {
            "type": "object",
            "properties": {
                "about": {
                    "description": "Информация о себе",
                    "type": "string"
                },
                "city": {
                    "description": "Город",
                    "type": "string"
                },
                "desired_position": {
                    "description": "Желаемая должность",
                    "type": "string"
                },
                "full_name": {
                    "description": "Полное имя",
                    "type": "string"
                },
                "parsed_data": {
                    "description": "Дополнительные данные, извлеченные из резюме"
                },
                "skills": {
                    "description": "Навыки",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user": {
                    "description": "Информация о пользователе",
                    "allOf": [
                        {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.UserResponse"
                        }
                    ]
                }
            }
        },
        "jumyste-app-backend_internal_dto.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "User registered successfully"
                }
            }
        },
...
        "jumyste-app-backend_internal_dto.UserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "Электронная почта",
                    "type": "string"
                },
                "first_name": {
                    "description": "Имя пользователя",
                    "type": "string"
                },
                "id": {
                    "description": "ID пользователя",
                    "type": "integer"
                },
                "last_name": {
                    "description": "Фамилия пользователя",
                    "type": "string"
                },
                "profile_picture": {
                    "description": "Профильное изображение",
                    "type": "string"
                },
                "role_id": {
                    "description": "ID роли пользователя",
                    "type": "integer"
                }
            }
        },
...
        "jumyste-app-backend_internal_entity.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "profile_picture": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}